move log class dir
